+++ packages/blocks/core/js/third-party/bootstraps.js
@@ -2,3 +2,4 @@
 
 export * from './outermost-icon-block/bootstrap';
 export * from './blocksy-breadcrumbs/bootstrap';
+export * from './blocksy-about-me/bootstrap';